Rename folder to migratableModelLoader
